requirements.txt
beautifulsoup4>=4.14.3
crewai>=1.6.1
fastapi>=0.129.0
fastmcp>=2.14.5
feedparser>=6.0.12
google-api-python-client>=2.190.0
google-auth>=2.48.0
google-auth-httplib2>=0.3.0
google-auth-oauthlib>=1.2.4
langchain>=1.2.10
langchain-groq>=1.1.2
litellm>=1.81.12
lxml-html-clean>=0.4.3
mcp>=1.26.0
newspaper3k>=0.2.8
readability-lxml>=0.8.4.1
requests>=2.32.5
streamlit>=1.54.0
uvicorn>=0.40.0